osd: Move ECListener to its own file.
This will be used by both legacy and optimised EC, so is move out to be common.
